- Added line chart examples

+++ composeApp/src/commonMain/kotlin/com/skellyapps/charts/line/model/ClickedPoint.kt
@@ -2,6 +2,6 @@ package com.skellyapps.charts.line.model
 
 internal data class ClickedPoint(
     val index: Int,
-    val lineTag: Byte,
+    val lineTag: Int,
     val isLeftAxis: Boolean
 )
