+++ composeApp/src/commonMain/kotlin/com/skellyapps/charts/line/model/DraggedPoint.kt
@@ -2,6 +2,6 @@ package com.skellyapps.charts.line.model
 
 internal data class DraggedPoint(
     val index: Int,
-    val lineTag: Byte,
+    val lineTag: Int,
     val isLeftAxis: Boolean
 )
